refactor: move ConnectTo to commands
